Update README.md
Tracking System using GPS and GSM Technologies

This project was developed as part of the requirements for a bachelor's degree course, specifically within a Project Course. Its core objective is to enable tracking of lost items by utilizing GPS and GSM technology with SIM Card integration. In the event of an item being lost and connected to the device, users can send an SMS to the GSM, which is linked to their phone via an app. This app, featuring a Google Map interface, allows users to effortlessly check and track the location of the lost item.